fix(test): refactor helpers into common files
Move unittest and functional test shared code into helper files
