+++ async_producer_test.go
@@ -19,55 +19,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-const TestMessage = "ABC THE MESSAGE"
-
-func closeProducerWithTimeout(t *testing.T, p AsyncProducer, timeout time.Duration) {
-	var wg sync.WaitGroup
-	p.AsyncClose()
-
-	closer := make(chan struct{})
-	timer := time.AfterFunc(timeout, func() {
-		t.Error("timeout")
-		close(closer)
-	})
-	defer timer.Stop()
-
-	wg.Add(2)
-	go func() {
-		defer wg.Done()
-		for {
-			select {
-			case <-closer:
-				return
-			case _, ok := <-p.Successes():
-				if !ok {
-					return
-				}
-				t.Error("Unexpected message on Successes()")
-			}
-		}
-	}()
-	go func() {
-		defer wg.Done()
-		for {
-			select {
-			case <-closer:
-				return
-			case msg, ok := <-p.Errors():
-				if !ok {
-					return
-				}
-				t.Error(msg.Err)
-			}
-		}
-	}()
-	wg.Wait()
-}
-
-func closeProducer(t *testing.T, p AsyncProducer) {
-	closeProducerWithTimeout(t, p, 5*time.Minute)
-}
-
 func expectResultsWithTimeout(t *testing.T, p AsyncProducer, successCount, errorCount int, timeout time.Duration) {
 	t.Helper()
 	expect := successCount + errorCount
@@ -1612,27 +1563,6 @@ func TestBrokerProducerShutdown(t *testing.T) {
 
 	_ = producer.Close()
 	mockBroker.Close()
-}
-
-type appendInterceptor struct {
-	i int
-}
-
-func (b *appendInterceptor) OnSend(msg *ProducerMessage) {
-	if b.i < 0 {
-		panic("hey, the interceptor has failed")
-	}
-	v, _ := msg.Value.Encode()
-	msg.Value = StringEncoder(string(v) + strconv.Itoa(b.i))
-	b.i++
-}
-
-func (b *appendInterceptor) OnConsume(msg *ConsumerMessage) {
-	if b.i < 0 {
-		panic("hey, the interceptor has failed")
-	}
-	msg.Value = []byte(string(msg.Value) + strconv.Itoa(b.i))
-	b.i++
 }
 
 func testProducerInterceptor(
