+++ client_test.go
@@ -14,14 +14,6 @@ import (
 	"github.com/rcrowley/go-metrics"
 )
 
-func safeClose(t testing.TB, c io.Closer) {
-	t.Helper()
-	err := c.Close()
-	if err != nil {
-		t.Error(err)
-	}
-}
-
 func TestSimpleClient(t *testing.T) {
 	seedBroker := NewMockBroker(t, 1)
