+++ encoder_decoder_fuzz_test.go
@@ -1,4 +1,4 @@
-//go:build go1.18
+//go:build go1.18 && !functional
 
 package sarama
